moved IndexRanges into separate file
